pyproject.toml
[build-system]
requires = ["setuptools>=64", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "cora"
version = "0.1.0"
description = "Utilities for building healthcare assistants on Vapi."
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "python-dotenv>=0.21",
    "vapi-server-sdk",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
